+++ lib/Sort/quickSort.h
@@ -5,8 +5,8 @@ using namespace std;
 
 namespace AEPKSS::Sort
 {
-    void quick_sort(vector<uint32_t> &in);
+    void quick_sort(vector<size_t> &in);
 }
 
-static void sort(vector<uint32_t> &in, int32_t left, int32_t right);
-static uint32_t partition(vector<uint32_t> &in, int32_t left, int32_t right);
+static void sort(vector<size_t> &in, size_t left, size_t right);
+static size_t partition(vector<size_t> &in, size_t left, size_t right);
